Fully initialize List state in constructor
The state of a List component was not fully initialized by the
constructor and relied upon the `componentDidMount` method to finish
state initialization, which was bad.  Not only should state not be
modified in `componentDidMount`, but the part of the state that was not
initialized in the constructor was the query criteria for fetching
granules.

This meant that _all_ granules were fetched and initially populated in
the List component, which caused the error granules list on the home
page to list _all_ granules, not only error granules.

The state is now fully initialized in the List constructor.

In addition, this commit includes a bit of additional cleanup and
refactoring throughout, including potential reduction of calls to
`this.setState` in `componentDidUpdate`, which should probably also not
exist.  This method should likely be revisited later to determine how
all calls to `this.setState` might be removed from it.

CUMULUS-1427