Fix unhandled rejection in `getMMTLinks` function
The `getMMTLinks` function in `app/scripts/actions/index.js` no longer
throws the following error when a collection does not contain values for
the properties `cmrProvider` and `cmrEnvironment`:

    Unhandled rejection TypeError: Cannot read property 'then' of null

This could be seen in the browser console after navigating to the
Executions page and then refreshing the page while running against the
local API server. This was due to the `getMMTLinkFromCmr` function
returning `null` due to a failure to populate the state with the Cumulus
Instance Metadata.

This fix is not specific to this Jira issue, but is some cleanup to
avoid creating yet another Jira issue.

CUMULUS-1308