Fix failed prop type error for checkboxes in tables
Initial rendering of checkboxes within tables, such as on the
Collections, Granules, and Rules pages, no longer results in the
following warning:

    Warning: Failed prop type: You provided a `checked` prop to a form
    field without an `onChange` handler. This will render a read-only
    field. If the field should be mutable use `defaultChecked`.
    Otherwise, set either `onChange` or `readOnly`.

Since the `tr` elements containing the checkboxes have `onChange`
handlers to manage selection/deselection of the checkboxes, the
checkboxes now have `readOnly` set on them.

This fix is not specific to this Jira issue, but is some cleanup to
avoid creating yet another Jira issue.

CUMULUS-1308