Fix invalid value for prop `className` on <a> tag
Rendering the sidebar no longer generates the following warning in
the browser console:

    Warning: Invalid value for prop `className` on <a> tag. Either
    remove it from the element, or pass a string or number value to keep
    it in the DOM.

This was caused by always assuming the value extracted for use as a
class name was a string.  However, it is sometimes a function, and in
such cases, the warning above was generated.  Now the extracted value
is used as a class name only when it is a string.

This fix is not specific to this Jira issue, but is some cleanup to
avoid creating yet another Jira issue.

CUMULUS-1308