Fix flaky Cypress integration test
The test for adding a new collection now reliably passes when it should.
Removed the part of the test that would navigate to the edit collection
page to check whether or not the new collection was indeed added.
Instead, a call to the API is made to attempt to retrieve the newly
added collection.  This proved to be a reliable solution, and is
arguably more appropriate anyway, as the goal of the test is to ensure
that the collection is persisted, not that it can be edited.

CUMULUS-1308