Pass full provider object to API on edit
When submitting an edit to a Provider, the entire object is now passed
to the API (via PUT) rather than a partial object (containing only the
changed properties).  This is in order to remain compatible with the API
changes that were implemented as part of this same Jira issue.  However,
this dashboard change does *not* depend upon the changes to the API, as
passing the full object to the unchanged API will also work as expected.
This dashboard change is necessary in order to avoid breakage once it is
used with the updated API.

Added a new integration test for this dashboard change, which required
adding 2 new dev dependencies: jsdom and jsdom-global.

CUMULUS-1308