Strengthen "add collection" cypress test
Added previously removed condition in the "add colletion" cypress test
to ensure that the newly added collection is shown on the collections
list page.  Further, moved mocking of CMR request responses out of the
"MMR" test and into `beforeEach` so that the "add collection" test
could also leverage the mocked responses.  This has eliminated the
previously flaky behavior of the "add collection" test, where it would
not consistently succeed.

CUMULUS-1308