CUMULUS-1308 handle deleted fields via api (#585)
* Ignore IntelliJ project files and yarn error log

CUMULUS-1427

* Add failing test

CUMULUS-1427

* Fully initialize List state in constructor

The state of a List component was not fully initialized by the
constructor and relied upon the `componentDidMount` method to finish
state initialization, which was bad.  Not only should state not be
modified in `componentDidMount`, but the part of the state that was not
initialized in the constructor was the query criteria for fetching
granules.

This meant that _all_ granules were fetched and initially populated in
the List component, which caused the error granules list on the home
page to list _all_ granules, not only error granules.

The state is now fully initialized in the List constructor.

In addition, this commit includes a bit of additional cleanup and
refactoring throughout, including potential reduction of calls to
`this.setState` in `componentDidUpdate`, which should probably also not
exist.  This method should likely be revisited later to determine how
all calls to `this.setState` might be removed from it.

CUMULUS-1427

* Fix unhandled rejection in `getMMTLinks` function

The `getMMTLinks` function in `app/scripts/actions/index.js` no longer
throws the following error when a collection does not contain values for
the properties `cmrProvider` and `cmrEnvironment`:

    Unhandled rejection TypeError: Cannot read property 'then' of null

This could be seen in the browser console after navigating to the
Executions page and then refreshing the page while running against the
local API server. This was due to the `getMMTLinkFromCmr` function
returning `null` due to a failure to populate the state with the Cumulus
Instance Metadata.

This fix is not specific to this Jira issue, but is some cleanup to
avoid creating yet another Jira issue.

CUMULUS-1308

* Fix failed prop type error for checkboxes in tables

Initial rendering of checkboxes within tables, such as on the
Collections, Granules, and Rules pages, no longer results in the
following warning:

    Warning: Failed prop type: You provided a `checked` prop to a form
    field without an `onChange` handler. This will render a read-only
    field. If the field should be mutable use `defaultChecked`.
    Otherwise, set either `onChange` or `readOnly`.

Since the `tr` elements containing the checkboxes have `onChange`
handlers to manage selection/deselection of the checkboxes, the
checkboxes now have `readOnly` set on them.

This fix is not specific to this Jira issue, but is some cleanup to
avoid creating yet another Jira issue.

CUMULUS-1308

* Fix invalid value for prop `className` on <a> tag

Rendering the sidebar no longer generates the following warning in
the browser console:

    Warning: Invalid value for prop `className` on <a> tag. Either
    remove it from the element, or pass a string or number value to keep
    it in the DOM.

This was caused by always assuming the value extracted for use as a
class name was a string.  However, it is sometimes a function, and in
such cases, the warning above was generated.  Now the extracted value
is used as a class name only when it is a string.

This fix is not specific to this Jira issue, but is some cleanup to
avoid creating yet another Jira issue.

CUMULUS-1308

* Fix flaky Cypress integration test

The test for adding a new collection now reliably passes when it should.
Removed the part of the test that would navigate to the edit collection
page to check whether or not the new collection was indeed added.
Instead, a call to the API is made to attempt to retrieve the newly
added collection.  This proved to be a reliable solution, and is
arguably more appropriate anyway, as the goal of the test is to ensure
that the collection is persisted, not that it can be edited.

CUMULUS-1308

* Eliminate "caniuse-lite is outdated" message during testing

Manually removed `caniuse-lite` entry from `yarn.lock`, ran
`yarn upgrade caniuse-lite`, then ran `yarn install` to eliminate the
following output at the start of running tests via `yarn test`:

    Browserslist: caniuse-lite is outdated. Please run next command
    `yarn upgrade`

CUMULUS-1308

* Upgrade Cypress to latest version (3.4.1)

CUMULUS-1308

* Pass full provider object to API on edit

When submitting an edit to a Provider, the entire object is now passed
to the API (via PUT) rather than a partial object (containing only the
changed properties).  This is in order to remain compatible with the API
changes that were implemented as part of this same Jira issue.  However,
this dashboard change does *not* depend upon the changes to the API, as
passing the full object to the unchanged API will also work as expected.
This dashboard change is necessary in order to avoid breakage once it is
used with the updated API.

Added a new integration test for this dashboard change, which required
adding 2 new dev dependencies: jsdom and jsdom-global.

CUMULUS-1308

* Update change log for CUMULUS-1308

* Strengthen "add collection" cypress test

Added previously removed condition in the "add colletion" cypress test
to ensure that the newly added collection is shown on the collections
list page.  Further, moved mocking of CMR request responses out of the
"MMR" test and into `beforeEach` so that the "add collection" test
could also leverage the mocked responses.  This has eliminated the
previously flaky behavior of the "add collection" test, where it would
not consistently succeed.

CUMULUS-1308