[CUMULUS-123] Fix filtering bug on List component
There was a bug introduced into the List component when trying
to filter on objects where queries weren't being sent to the
Cumulus API correctly. For example, instead of querying for
"state=completed" in the Granules table, it would query for
"params[state]=completed" causing the API to throw an error.
This changes fixes it so the params hash is expanded correctly
in the query object, fixing the bug.