Cumulus 123 mhs (#594)
* [CUMULUS-123] Fix filtering bug on List component

There was a bug introduced into the List component when trying
to filter on objects where queries weren't being sent to the
Cumulus API correctly. For example, instead of querying for
"state=completed" in the Granules table, it would query for
"params[state]=completed" causing the API to throw an error.
This changes fixes it so the params hash is expanded correctly
in the query object, fixing the bug.

* Adds update history for select and certain onchange events.

WIP commit.  Need to examine search box.

* Add URL history to dropdowns.

From what I read, if you use react-router with redux, router is supposed to be
the URL truth and store had the state truth.

* Change statusToLabelFactory into a regular function.

This isn't 100%.  But I wasn't updating the value correctly on reload for a
component that was loaded from scratch from the URL query parameters

* Code cleanup.

* Adds cypress tests for URL and dropdowns

* Updates react to 16.8 for Hooks

Add Lint options for react-hooks.

Changed name of two imported functions because hooks are presumed to begin with
useFunction.

* Upgrade Ava to 2.4.0; pin cmrjs 1.11.0

Upgrades all minor versions.

* Upgrade react and cypress/browserify-preprocessor

* Migrate from @cumulus/cmrjs to @cumulus/cmr-client

* Roll back upgrade of cmrjs to cmr-client

We have a ticket to fix.
 https://bugs.earthdata.nasa.gov/browse/CUMULUS-1624

* This works before refactoring.

* Move dispatch out of constructor.

This duplicates code because we have to use both lifecycle methods
componentDidMount and componentDidUpdate

* Adds cypress tests for search components.

* Finish those tests.

* Refactor url / location / browser

Rather than put it as a reducer, moslty because I wasn't sure how to access the
browser from one, I'm importing common code between the search and dropdown components.

* Adds tests for url-helper

* reformat JSX

remove extra console.log.

* Select failed by typing, instead of depending on order.

Old test tested runing as first clicked item, now we use the autocomplete to
select failed.

* PR feedback update.

* Removes invalid url test.

This behavior is fine for now, but may not be the desired behavior in the
future.  So remove the test so that is doesn't fail if the behavior is changed.