Change enviroment variable for fakeapi. Update test ports. (#617)
Before, `PORT` was used as both the env var for overriding the fake api
port (normally 5001) as well as the dashboard server port (normally 3000).

This changes the envvar to FAKEAPIPORT and updates the tests to select the FAKEAPIPORT
rather than being hard coded to 5001.  This lets us run two version of the
dashboard at once during testing.