Goto granules on delete of collection with granules
When clicking Delete on collection overview page, new modal sequence is
shown, including prompting user when collection has associated granules.

In addition, changed test data such that a rule is associated with
collection MOD09GQ instead of MOD09GK. Otherwise test coverage is not
possible because MOD09GK has granules associated with it. By associating
a rule with one of them and granules with the other, we can properly
test deleting a collection with a rule as well as test deleting a
collection with granules.

CUMULUS-1520