Hard Coded waits.
I have added 4 sec to the run time by putting in hard coded cy.waits.

I think what is happening is that the DELETE is returning successful but that
the containers elastic/dynamo are not updated when the GET /collections/limit=*
is run, and we still see 5 collections.

Although, there is still the weirdness with the stuff not being visible in the
dom, but still visible to our eyes.