Add onChange property to Datepicker component.
This is a callback that will be called when the datepicker is updated.

It's called at the end of dispatchDropdownUpdate (Which is called when the
dropdown is selected or the datepicker cleared.)

It's also called at the end of handleDateTimeRangeChange, which will trigger
when a date is selected or typed into one of the datetime boxes.