Configurable history type.
Adds optional Environmental Varaible to allow a user to deploy the dashboard to
be served by Cumulus API's dashboard endpoint.

To do so set SERVED_BY_CUMULUS_API to true and build the dashboard.
This change is allowed by using a hashrouter so that you don't go back to the
server for anything.  Which is what we need to serve the entire application
from a single endpoint.