Makes errors smarter. Adds tests for middleware.
Errors now will use the message from the respose.data field unless that is
falsey, in which case it will grab an available message from
error.message. This allows us to keep the existing behavior for most errors,
but get information from the errors for a misconfigured ES stack (e.g. wrong
name or password)