Updates Kibana links to remove queries that include _index
This also changes the path to the currently resolved base path.
before:/app/kibana#/discover?
after:/app/discover#/
This is just an update to the kibana version. the old paths worked, but get
resolved to the new paths.

I generated the new values by checking the old links in kibana, removing the
query filter for _index, and making sure the same data was displayed for values
that had data.  Following this same method, I updated the links for the values
that did not have data.

I have not updated the s3 links yet because there appears to be actual errors
in them at the moment.