Update unpublished granule fixtures to explicitly have cmrLink 'null'
This addresses an issue with the modified test setup factory where
granules are being created with a value for cmrLink despite being
unpublished.   This shouldn't happen in Core -- update granule display
to validate granule is published AND has a link before making a link,
update fixtures to explicitly set cmrLink to null as well.