Update axe invocation to catch/log error instead of throwing
Axe isn't being used in formal tests, but the logging is useful during
development to identify issues.    Unfortunately cypress tests are
importing index repeatedly and axe is throwing and 'already running'
exception that isn't valueable in the context of the cypress tests.
This update logs, but does not cause an exception to prevent cypress
from failing the test.