Update dashboard to work with Core v16 release (#1103)
* WIP commit, see ticket for detail

Contains work toward 3318 objective, and fixing Cypress test runs in
combination with 3318 WIP ticket on Core.    Also has work related to
12.0.1 release.

* remove comment

* Update webback to fix security vulnerability

* add lock file

* bump cumulus versions

* Replace ssh url with http to fix CI build failure

* Use new alpha API version

* update package.lock

* update to https

* update core alpha version

* Update unpublished granule fixtures to explicitly have cmrLink 'null'

This addresses an issue with the modified test setup factory where
granules are being created with a value for cmrLink despite being
unpublished.   This shouldn't happen in Core -- update granule display
to validate granule is published AND has a link before making a link,
update fixtures to explicitly set cmrLink to null as well.

* Fix lint

* Fix CHANGELOG

* Update API compat configuration

* Update axe invocation to catch/log error instead of throwing

Axe isn't being used in formal tests, but the logging is useful during
development to identify issues.    Unfortunately cypress tests are
importing index repeatedly and axe is throwing and 'already running'
exception that isn't valueable in the context of the cypress tests.
This update logs, but does not cause an exception to prevent cypress
from failing the test.

* bump api version

* Fix rules spec error text

* Add axe abort for cypress

* Update index to correctly use env var to determine axe load

* Disable `axe` due to cypress incompatibility

* Fix provider spec title, regression to be addressed in CUMULUS-3380

* Remove unused argument

---------

Co-authored-by: Nate Pauzenga <npauzenga@gmail.com>